move allow list role to different file
